Added: Runge-Kutta and Error checks
Runge-Kutta method added as per reference material- I used the Runge-Kutta videos by LearnChemE on Youtube to learn about RK method

Error checks for dynamic model convergence and tolerance were added. In addition to the reference model I added an error check for temperatures not-converging through miscalculation by stopping the model when the values are calculated outside of the max/min initial temperatures.

I next plan to work on creating a UI to allow user to more easily set-up and control the model.